DT-835 Convert institutional report controller tests to unit tests
