+++ src/main/java/uk/gov/justice/digital/delius/data/api/InstitutionalReport.java
@@ -1,5 +1,6 @@
 package uk.gov.justice.digital.delius.data.api;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -17,6 +18,7 @@ public class InstitutionalReport {
     private Long offenderId;
     private Conviction conviction;
 
+    @JsonIgnore
     @ApiModelProperty(notes = "Deprecated - Use conviction to access sentence")
     public Sentence getSentence() {
         return Optional.ofNullable(conviction).map(Conviction::getSentence).orElse(null);
